move all passwords list logic inside the component, not the home page
